pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hyperliquid-mcp"
version = "0.1.0"
description = "MCP server for Hyperliquid — control your perps from Claude"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["hyperliquid", "mcp", "claude", "trading", "defi", "perpetuals"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "mcp>=1.2.0",
    "hyperliquid-python-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "eth-account>=0.10.0",
]

[project.scripts]
hyperliquid-mcp = "hyperliquid_mcp.server:main"

[project.urls]
Homepage = "https://github.com/kikerub/hyperliquid-mcp"
Repository = "https://github.com/kikerub/hyperliquid-mcp"
Issues = "https://github.com/kikerub/hyperliquid-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/hyperliquid_mcp"]
